requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
sqlalchemy
PyMySQL
redis
pymilvus
python-jose[cryptography]
passlib
python-multipart
httpx
pypdf
PyMuPDF
rapidocr-onnxruntime
Pillow
pytest
